Safari-611.1.10/LayoutTests/editing/selection/ios/hide-selection-in-non-focused-element-expected.txt

Hello world

Verifies that text selection is suppressed when the selected content is in a hidden editable root, even when focusing a different element. To manually run the test, tap the editable area in the red box, and then tap the button below it. Platform selection views should not become visible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS selectionRects.length is 0
PASS getSelection().toString() is "Hello world"
PASS document.activeElement is focusTarget
PASS successfullyParsed is true

TEST COMPLETE
